29Items-Projects/3Python/1E-Commerce Platform/backend/requirements.txt
# Django & REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Database
psycopg[binary]>=3.1,<4.0
SQLAlchemy>=2.0,<3.0

# Async Task Processing
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
django-celery-results>=2.5,<3.0

# Search
elasticsearch>=8.11,<9.0
elasticsearch-dsl>=8.11,<9.0
django-elasticsearch-dsl>=8.0,<9.0

# ML & Data Processing
numpy>=1.26,<2.0
torch>=2.1,<3.0
scikit-learn>=1.3,<2.0
pandas>=2.1,<3.0

# Storage
boto3>=1.34,<2.0
django-storages>=1.14,<2.0
Pillow>=10.1,<11.0

# Payments
stripe>=7.0,<8.0

# Utilities
python-decouple>=3.8,<4.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0

# Monitoring
sentry-sdk>=1.38,<2.0
structlog>=23.2,<24.0
python-json-logger>=2.0,<3.0

# Caching
django-redis>=5.4,<6.0
